% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBifExample.R
\name{deBifExample}
\alias{deBifExample}
\title{Examples of phaseplane analysis of a system of ODEs}
\usage{
deBifExample(example)
}
\arguments{
\item{example}{(string, optional)
\preformatted{}
              Name of the example. If not provided a list of examples
              is returned}
}
\value{
None.
}
\description{
\code{deBifExample}
}
\details{
deBifExample(example)


Function runs one of the examples provided with the deBif package
}
