% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBifHelp.R
\name{deBifHelp}
\alias{deBifHelp}
\title{Opens the deBif manual}
\usage{
deBifHelp()
}
\value{
None.
}
\description{
\code{deBifHelp} opens the manual of the the deBif package in html format.
}
\examples{
if(interactive()){
deBifHelp()
}

}
