% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deBifReset.R
\name{deBifReset}
\alias{deBifReset}
\title{Reloads the deBif package}
\usage{
deBifReset()
}
\value{
None.
}
\description{
\code{deBifReset} unloads and reloads the deBif package.
}
\examples{
if(interactive()){
deBifReset()
}

}
