% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runExample.R
\name{runExample}
\alias{runExample}
\title{Examples of phaseplane analysis of a system of ODEs}
\usage{
runExample(example)
}
\arguments{
\item{example}{(string, optional)
\preformatted{}
              Name of the example. If not provided a list of examples
              is returned}
}
\description{
\code{runExample}
}
\details{
runExample(example)

Function runs one of the examples provided with the deBif package
}
